+++ modules/Student_Billing/includes/DailyTotals.php
@@ -17,8 +17,8 @@ DrawHeader( _programMenu( 'totals' ) );
 echo '<form action="' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&program=totals'  ) . '" method="GET">';
 
 DrawHeader( _( 'Report Timeframe' ) . ': ' .
-	PrepareDate( $start_date, '_start' ) . ' ' . _( 'to' ) . ' ' .
-	PrepareDate( $end_date, '_end' ), SubmitButton( _( 'Go' ) ) );
+	PrepareDate( $start_date, '_start', false ) . ' ' . _( 'to' ) . ' ' .
+	PrepareDate( $end_date, '_end', false ) . ' ' . SubmitButton( _( 'Go' ) ) );
 
 echo '</form>';